+++ src/js/Utils/createNewCard.js
@@ -2,13 +2,14 @@ import Card from "../../js/Components/Card";
 import Api from "../Api/MainApi";
 import monthFormat from "./monthFormat";
 
+const container = document.querySelector('.results__container') || null;
 const serverUrl = 'https://api.explorenews.gq';
 const api = new Api({
   baseUrl: serverUrl
 })
 
 // Создание экземпляра карточки со статьей
-export default function createNewCard(keyword, img, link, date, title, text, source, isLoggedIn){
+export default function createNewCard(keyword, img, link, date, title, text, source, isLoggedIn, cardId){
   return new Card({
     keyword,
     img,
@@ -18,6 +19,8 @@ export default function createNewCard(keyword, img, link, date, title, text, sou
     text,
     source,
     isLoggedIn,
+    cardId,
+    container,
     monthFormat,
     saveArticle: (keyword, title, text, date, source, link, image)=> {
       return api.saveArticle(keyword, title, text, date, source, link, image);
